requirements.txt
asgiref==3.3.0
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.3
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
Pygments==2.7.2
pytz==2020.4
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
rsa~=4.6
pycryptodome~=3.9.9
pycryptodomex~=3.9.9

# https://aka.ms/vs/16/release/vc_redist.x64.exe get Microsoft Visual C++ 14.0 from here for Windows users whom trying to install pycrypto package


